+++ src/casts.ts
@@ -7,7 +7,12 @@ export function int_from_bytes(b: Bytes|None): int {
   if(!b || b.length === 0){
     return 0;
   }
-  return parseInt(b.hex(), 16);
+  const unsigned32 = parseInt(b.hex(), 16);
+  // If the first bit is 1, it is recognized as a negative number.
+  if(b.get_byte_at(0) & 0x80){
+    return unsigned32 - (1 << (b.length*8));
+  }
+  return unsigned32;
 }
 
 export function int_to_bytes(v: int): Bytes {